feat: PgService corrected
